requirements.txt
python-telegram-bot==21.6
openai==1.51.0
supabase==2.7.4
faiss-cpu==1.8.0
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
pypdf==4.3.1
python-docx==1.1.2
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0
pillow==10.4.0
pydantic==2.9.2
numpy==1.26.4
tiktoken==0.7.0
